+++ src/main/java/com/example/practice/robot/RobotRepository.java
@@ -2,5 +2,5 @@ package com.example.practice.robot;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RobotRepository extends JpaRepository<Robot, Long> {
+public interface RobotRepository extends JpaRepository<Robot, Integer> {
 }
